refactor: move content view filters and filter rules into own Go file
